Fixed bug with setting final position of animation. Also made sure to set a begin time on keyframe animations so we get only one run (bug was repeating). Also confirmed we can parallel multiple property changes.

+++ Rouse/Enums/KeyPaths.cs
@@ -7,7 +7,7 @@ namespace RouseLib
 	{
 		public float? PositionX;
 		public float? PositionY;
-		public float? Opacity;
+		public double? Opacity;
 		public float? Rotation;
 		public float? RotationX;
 		public float? RotationY;
